Make bootstrap.sh idempontent

+++ bootstrap.sh
@@ -22,7 +22,7 @@ fi
 for c in config/*; do
     link_config "$c"
 done
-ln -s "$PWD/tmux.conf" "$HOME/.tmux.conf"
+test -f "$HOME/.tmux.conf" || ln -s "$PWD/tmux.conf" "$HOME/.tmux.conf"
 
 for c in dynamic_config/*; do
     cp -r "$c" ~/.config
